miden_vm::advice

Struct AdviceMap

pub struct AdviceMap(/* private fields */);

Defines a set of non-deterministic (advice) inputs which the VM can access by their keys.

Each key maps to one or more field element. To access the elements, the VM can move the values associated with a given key onto the advice stack using adv.push_mapval instruction. The VM can also insert new values into the advice map during execution.

Implementations

impl AdviceMap

pub fn get(&self, key: &Word) -> Option<&Arc<[Felt]>>

Returns the values associated with given key.

pub fn contains_key(&self, key: &Word) -> bool

Returns true if the key has a corresponding value in the map.

pub fn insert( &mut self, key: Word, value: impl Into<Arc<[Felt]>>, ) -> Option<Arc<[Felt]>>

Inserts a value, returning the previous value if the key was already set.

pub fn remove(&mut self, key: &Word) -> Option<Arc<[Felt]>>

Removes the value associated with the key and returns the removed element.

pub fn iter(&self) -> impl Iterator<Item = (&Word, &Arc<[Felt]>)>

Return an iteration over all entries in the map.

pub fn len(&self) -> usize

Returns the number of key value pairs in the advice map.

pub fn is_empty(&self) -> bool

Returns true if the advice map is empty.

pub fn entry(&mut self, key: Word) -> Entry<'_, Word, Arc<[Felt]>>

Gets the given key’s corresponding entry in the map for in-place manipulation.

pub fn merge( &mut self, other: &AdviceMap, ) -> Result<(), ((Word, Arc<[Felt]>), Arc<[Felt]>)>

Merges all entries from the given AdviceMap into the current advice map.

If an entry from the new map already exists with the same key but different value, an error is returned containing the existing entry along with the value that would replace it. The current map remains unchanged.