miden_vm::crypto::dsa::falcon512_poseidon2

Struct Polynomial

pub struct Polynomial<F> {
    pub coefficients: Vec<F>,
}

Represents a polynomial with coefficients of type F.

Fields
coefficients: Vec<F>

Coefficients of the polynomial, ordered from lowest to highest degree.

Implementations

impl<F> Polynomial<F>
where F: Clone,

pub fn new(coefficients: Vec<F>) -> Polynomial<F>

Creates a new polynomial from the provided coefficients.

impl<F> Polynomial<F>
where F: Sub<Output = F> + Mul<Output = F> + AddAssign + Zero + Div<Output = F> + Clone + Inverse,

pub fn hadamard_mul(&self, other: &Polynomial<F>) -> Polynomial<F>

Multiplies two polynomials coefficient-wise (Hadamard multiplication).

pub fn hadamard_div(&self, other: &Polynomial<F>) -> Polynomial<F>

Divides two polynomials coefficient-wise (Hadamard division).

pub fn hadamard_inv(&self) -> Polynomial<F>

Computes the coefficient-wise inverse (Hadamard inverse).

impl<F> Polynomial<F>
where F: Zero + PartialEq + Clone,

pub fn degree(&self) -> Option<usize>

Returns the degree of the polynomial.

pub fn lc(&self) -> F

Returns the leading coefficient of the polynomial.

impl<F> Polynomial<F>
where F: Clone + Neg<Output = F> + MulAssign + AddAssign + Div<Output = F> + One + Sub<Output = F> + Zero + PartialEq,

The following implementations are specific to cyclotomic polynomial rings, i.e., F[ X ] / <X^n + 1>, and are used extensively in Falcon.

pub fn reduce_by_cyclotomic(&self, n: usize) -> Polynomial<F>

Reduce the polynomial by X^n + 1.

pub fn field_norm(&self) -> Polynomial<F>

Computes the field norm of the polynomial as an element of the cyclotomic ring F[ X ] / <X^n + 1 > relative to one of half the size, i.e., F[ X ] / <X^(n/2) + 1> .

Corresponds to formula 3.25 in the spec [1, p.30].

pub fn lift_next_cyclotomic(&self) -> Polynomial<F>

Lifts an element from a cyclotomic polynomial ring to one of double the size.

pub fn galois_adjoint(&self) -> Polynomial<F>

Computes the galois adjoint of the polynomial in the cyclotomic ring F[ X ] / < X^n + 1 > , which corresponds to f(x^2).

impl<F> Polynomial<F>
where F: Sub<Output = F> + Mul<Output = F> + AddAssign + Zero + Div<Output = F> + Clone,

pub fn karatsuba(&self, other: &Polynomial<F>) -> Polynomial<F>

Multiply two polynomials using Karatsuba’s divide-and-conquer algorithm.

impl<F> Polynomial<F>
where F: Zero + Clone,

pub fn shift(&self, shamt: usize) -> Polynomial<F>

Shifts the polynomial by the specified amount (adds leading zeros).

pub fn constant(f: F) -> Polynomial<F>

Creates a constant polynomial with a single coefficient.

pub fn map<G, C>(&self, closure: C) -> Polynomial<G>
where G: Clone, C: FnMut(&F) -> G,

Applies a function to each coefficient and returns a new polynomial.

pub fn fold<G, C>(&self, initial_value: G, closure: C) -> G
where C: FnMut(G, &F) -> G + Clone,

Folds the coefficients using the provided function and initial value.

impl Polynomial<FalconFelt>

pub fn norm_squared(&self) -> u64

Computes the squared L2 norm of the polynomial.

pub fn to_elements(&self) -> Vec<Felt>

Returns the coefficients of this polynomial as field elements.

pub fn mul_modulo_p( a: &Polynomial<FalconFelt>, b: &Polynomial<FalconFelt>, ) -> [u64; 1024]

Multiplies two polynomials over Z_p[x] without reducing modulo p. Given that the degrees of the input polynomials are less than 512 and their coefficients are less than the modulus q equal to 12289, the resulting product polynomial is guaranteed to have coefficients less than the Miden prime.

Note that this multiplication is not over Z_p[x]/(phi).

pub fn reduce_negacyclic(a: &[u64; 1024]) -> Polynomial<FalconFelt>

Reduces a polynomial, that is the product of two polynomials over Z_p[x], modulo the irreducible polynomial phi. This results in an element in Z_p[x]/(phi).

impl Polynomial<Felt>

pub fn to_elements(&self) -> Vec<Felt>

Returns the coefficients of this polynomial as Miden field elements.

impl Polynomial<i16>

pub fn to_balanced_values(&self) -> Vec<i16>

Returns the balanced values of the coefficients of this polynomial.