miden_vm

Module internal

Modules

event

Events represent single points in time during the execution of a program.

instrument

Attach a span to a std::future::Future.

Macros

event

Constructs a new Event.

Structs

InputFile

Input file struct that is used to deserialize input data from file. It consists of four components:

Level

Describes the level of verbosity of a span or event.

Enums

MerkleData

Struct used to deserialize merkle data from input file. Merkle data can be represented as a merkle tree or a Sparse Merkle Tree.

Attribute Macros

instrument

Instruments a function to create and enter a tracing span every time the function is called.