miden_vm

Module trace

Modules

chiplets

Structs

AuxTraceBuilders

ChipletsLengths

Contains trace lengths of all chiplets: hash, bitwise, memory, ACE, and kernel ROM.

ExecutionTrace

Execution trace which is generated when a program is executed on the VM.

RowIndex

A newtype wrapper around a usize value representing a step in the execution trace.

TraceBuildInputs

Inputs required to build an execution trace from pre-executed data.

TraceGenerationContext

TraceLenSummary

Contains the data about lengths of the trace parts.

Constants

CORE_TRACE_WIDTH

Functions

build_trace

Builds the main trace from the provided trace states in parallel.

build_trace_with_max_len

Same as build_trace, but with a custom hard cap.